started adding sanitization, finished cleaning up members triad

+++ html/application/controllers/member.controller.php
@@ -9,28 +9,44 @@ if(!class_exists('Member_Controller')) {
             parent::__construct($model);
         }
 
-        public function getUserId() {
-            return $this->model->getUserId($_SESSION['username']);
+        public function setCurrentList($list_title) {
+            $list_title = Utils::washInput($list_title);
+            if(!empty($list_title) && $this->model->listExists($list_title)) {
+                return $list_title;
+            } else {
+                return NULL;
+            }
         }
 
         public function getLists($user_id) {
             return $this->model->getLists($user_id);
         }
 
+        public function getCurrentUserId() {
+            return $this->model->getUserId($_SESSION['username']);
+        }
+
         public function addItem($list_title, $item_name) {
-            if(!empty($item_name)) {
-                $this->model->insertItem($list_title, $item_name);
+            $list_title = Utils::washInput($list_title);
+            $item_name = Utils::washInput($item_name);
+            $list_id = $this->model->getListId($list_title);
+            if(!empty($item_name) && !empty($list_id)) {
+                $this->model->insertItem($list_id, $item_name);
             }
         }
 
         public function removeItem($list_title, $item_name) {
+            $list_title = Utils::washInput($list_title);
+            $item_name = Utils::washInput($item_name);
             $list_id = $this->model->getListId($list_title);
-            if(!empty($item_name)) {
+            if(!empty($item_name) && !empty($list_id)) {
                 $this->model->deleteItem($list_id, $item_name);
             }
         }
 
         public function toggleItem($list_title, $item_name) {
+            $list_title = Utils::washInput($list_title);
+            $item_name = Utils::washInput($item_name);
             $list_id = $this->model->getListId($list_title);
             $this->model->tapItem($list_id, $item_name);
         }
